+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `League\Uri` will be documented in this file
 ### Added
 
 - `Component::getContent` returns the raw representation of a component
+- `User::getValue`, `Pass:getValue` and `Fragment::getValue` to access the decoded string representation
 - Support for PHP's magic methods `__debugInfo` and `__set_state` to improve debugging
 - Uri Modifier `Relativize` to complement `Resolve`
 
@@ -14,9 +15,9 @@ All Notable changes to `League\Uri` will be documented in this file
 
 - Components encoding/decoding to be more RFC3986 compliant
 - `Host::isPublicSuffixValid` when no `publicSuffix` information is found
-- `Host::getRegisterableDomain` must always return a string as per the Host Interface
-- `Host::getSubdomain` must always return a string as per the Host Interface
-- `Host::isPublicSuffixValid` must always return a string as per the Host Interface
+- `Host::getRegisterableDomain` must always return a string as per the host interface expected return type
+- `Host::getSubdomain` must always return a string as per the host interface expected return type
+- `Host::isPublicSuffixValid` must always return a string as per the host interface expected return type
 - On instantiation, query and fragment delimiter are preserved
 - `createFromComponents` failing with `parse_url` result when no path is defined
 - On URI transformation `InvalidArgumentException` exceptions are emitted instead of `RuntimeException` ones to normalize exception to PSR-7
@@ -24,6 +25,7 @@ All Notable changes to `League\Uri` will be documented in this file
 ### Deprecated
 
 - `Host::getLiteral` you should use `Host::getContent` instead
+- `Port::toInt` you should use `Port::getContent` instead
 
 ### Remove